Recursively traverse a user's follower network.
User.traverse takes a username and block as an argument. It finds the user's
followers, yielding each one to the block. It then repeats this process for
each of the followers, recursively.

This begs for a recursive function, but I kept exhausting the stack that way,
so this implementation will have to do.